fix: Now creates the zsh directory necessary to the mv operation

+++ utils.sh
@@ -142,8 +142,17 @@ zsh_configurations()
     fi
 
     # ------ TERMINAL KEYBINDINGS ------ #
-    # Moving keybindings file
-    mv ~/linuxtools/dotfiles/zsh-keybindings ~/usr/share/zsh/
+    # Creates zsh config directory
+    if [[ ! -d ~/usr/share/zsh ]]; then
+        mkdir ~/usr/share/zsh
+    fi
+
+    # Moves the keybindings dotfile
+    if [[ -f ~/linuxtools/dotfiles/zsh-keybindings ]]; then
+        mv ~/linuxtools/dotfiles/zsh-keybindings ~/usr/share/zsh/
+    else
+        echo "Source file not found."
+    fi
 
     # Sourcing keybindings
     cat << 'EOF' >> ~/.zshrc
